+++ torch_frame/data/mapper.py
@@ -1,7 +1,8 @@
 from __future__ import annotations
 
 from abc import ABC, abstractmethod
-from typing import Any, Callable, Iterable, Mapping
+from collections.abc import Iterable, Mapping
+from typing import Any, Callable
 
 import numpy as np
 import pandas as pd