+++ torch_frame/data/tensor_frame.py
@@ -69,10 +69,12 @@ class TensorFrame:
         feat_dict: dict[torch_frame.stype, TensorData],
         col_names_dict: dict[torch_frame.stype, list[str]],
         y: Tensor | None = None,
+        num_rows: int | None = None,
     ) -> None:
         self.feat_dict = feat_dict
         self.col_names_dict = col_names_dict
         self.y = y
+        self._num_rows = num_rows
         self.validate()
 
         # Quick mapping from column names into their (stype, idx) pairs in
@@ -93,7 +95,13 @@ class TensorFrame:
         num_rows = self.num_rows
         empty_stypes: list[torch_frame.stype] = []
         for stype_name, feats in self.feat_dict.items():
-            num_cols = len(self.col_names_dict[stype_name])
+            col_names = self.col_names_dict[stype_name]
+            if not isinstance(col_names, list):
+                raise ValueError(
+                    f"col_names_dict[{stype_name}] must be a list of column "
+                    f"names.")
+
+            num_cols = len(col_names)
             if num_cols == 0:
                 empty_stypes.append(stype_name)
 
@@ -175,6 +183,8 @@ class TensorFrame:
     @property
     def num_rows(self) -> int:
         r"""The number of rows in the :class:`TensorFrame`."""
+        if self._num_rows is not None:
+            return self._num_rows
         if self.is_empty:
             return 0
         feat = next(iter(self.feat_dict.values()))
@@ -293,7 +303,13 @@ class TensorFrame:
                 return x[index]
             return y
 
-        return self._apply(fn)
+        out = self._apply(fn)
+
+        if self._num_rows is not None:
+            dummy = torch.empty((self.num_rows, 0), device=self.device)
+            out._num_rows = dummy[index].size(0)
+
+        return out
 
     def __copy__(self) -> TensorFrame:
         out = self.__class__.__new__(self.__class__)